blueprint split of history
