+++ arm/ui/__init__.py
@@ -63,10 +63,12 @@ from arm.ui.settings.settings import route_settings  # noqa: E402,F811
 from arm.ui.logs.logs import route_logs  # noqa: E402,F811
 from arm.ui.auth.auth import route_auth  # noqa: E402,F811
 from arm.ui.database.database import route_database  # noqa: E402,F811
+from arm.ui.history.history import route_history  # noqa: E402,F811
 app.register_blueprint(route_settings)
 app.register_blueprint(route_logs)
 app.register_blueprint(route_auth)
 app.register_blueprint(route_database)
+app.register_blueprint(route_history)
 
 # Remove GET/page loads from logging
 import logging  # noqa: E402,F811
