+++ arm/ui/history/__init__.py
@@ -1,0 +1,3 @@
+"""
+ARM route blueprint for history pages
+"""
