+++ arm/ui/history/templates/history.html
@@ -53,7 +53,7 @@
                         </div>
                     </div>
                     <br>
-                    {% with pages=pages, page_name="history" %}
+                    {% with pages=pages, page_name="route_history.history" %}
                         {% include "pagination.html" %}
                     {% endwith %}
                     <!-- Main Body -->
@@ -88,7 +88,7 @@
                             </tbody>
                         </table>
                     </div>
-                    {% with pages=pages, page_name="history" %}
+                    {% with pages=pages, page_name="route_history.history" %}
                         {% include "pagination.html" %}
                     {% endwith %}
                 </div>
